requirements.txt
beautifulsoup4==4.14.3
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
curl_cffi==0.15.0
frozendict==2.4.7
idna==3.13
lxml==6.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
multitasking==0.0.13
numpy==2.4.4
pandas==3.0.2
pandas-datareader==0.10.0
peewee==4.0.5
platformdirs==4.9.6
protobuf==7.34.1
psycopg2-binary==2.9.12
pycparser==3.0
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
rich==15.0.0
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
yfinance==1.3.0
